requirements.txt
alembic==1.0.6
atomicwrites==1.3.0
attrs==18.2.0
Click==7.0
croniter==0.3.27
factory-boy==2.11.1
Faker==1.0.2
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
more-itertools==5.0.0
passlib==1.7.1
pluggy==0.8.1
psycopg2-binary==2.7.6.1
py==1.7.0
pytest==4.2.0
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
python-http-client==3.1.0
six==1.12.0
SQLAlchemy==1.2.16
text-unidecode==1.2
urllib3==1.24.1
Werkzeug==0.14.1
